refactor: remove dead code and reduce duplication across codebase
- Remove temporary _debug_log instrumentation with hardcoded local paths
  from outbound_telemetry/__init__.py and logging.py
- Simplify should_continue_investigation() routing logic: remove redundant
  boolean expression and replace print() calls with debug_print
- Extract shared make_boto3_client() factory to env.py; use in s3, cloudwatch,
  lambda clients (removes repeated boto3.client + region pattern)
- Extract temp_env context manager from 5 test files into shared conftest.py
- Extract infrastructure_available() from 3 test conftest files into shared infra.py
- Unify memory context loading in frame_problem.py to match root_cause_diagnosis pattern
- Remove duplicate import json in benchmark, unused tool import in cloudwatch client
