+++ app/memories/2026-02-19-test_pipeline-prior002.md
@@ -1,4 +1,4 @@
-# Session: 2026-02-19 09:05:12 UTC
+# Session: 2026-02-19 11:20:05 UTC
 
 - **Pipeline**: test_pipeline
 - **Alert ID**: prior002
