+++ app/outbound_telemetry/logging.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import json
 import logging
 import sys
-import time
 
 from opentelemetry import trace
 
@@ -12,27 +11,6 @@ from config.grafana_config import (
     get_otel_exporter_otlp_protocol,
     parse_otel_headers,
 )
-
-_DEBUG_LOG_PATH = "/Users/janvincentfranciszek/tracer-agent-2026/.cursor/debug.log"
-_DEBUG_SESSION_ID = "debug-session"
-_DEBUG_RUN_ID = "pre-fix"
-
-
-def _debug_log(hypothesis_id: str, location: str, message: str, data: dict) -> None:
-    payload = {
-        "sessionId": _DEBUG_SESSION_ID,
-        "runId": _DEBUG_RUN_ID,
-        "hypothesisId": hypothesis_id,
-        "location": location,
-        "message": message,
-        "data": data,
-        "timestamp": int(time.time() * 1000),
-    }
-    try:
-        with open(_DEBUG_LOG_PATH, "a", encoding="utf-8") as handle:
-            handle.write(json.dumps(payload) + "\n")
-    except Exception:
-        return
 
 
 def _get_log_exporter():
@@ -106,19 +84,6 @@ def setup_logging(resource) -> None:
     logs_endpoint = endpoint
     if protocol in ("http/protobuf", "http") and endpoint and not endpoint.endswith("/v1/logs"):
         logs_endpoint = endpoint.rstrip("/") + "/v1/logs"
-    # region agent log
-    _debug_log(
-        "H3",
-        "app/outbound_telemetry/logging.py:setup_logging",
-        "log_exporter_status",
-        {
-            "exporter": exporter.__class__.__name__ if exporter else None,
-            "protocol": protocol,
-            "endpoint": endpoint,
-            "logs_endpoint": logs_endpoint if exporter else None,
-        },
-    )
-    # endregion agent log
     if exporter is None:
         logging.getLogger(__name__).warning("OTLP log exporter is unavailable")
         return
