+++ tests/outbound_telemetry/logging_test.py
@@ -1,6 +1,4 @@
 import logging
-import os
-from contextlib import contextmanager
 
 from opentelemetry.sdk.resources import Resource
 
@@ -9,17 +7,7 @@ from app.outbound_telemetry.logging import (
     ensure_otel_logging,
     setup_logging,
 )
-
-
-@contextmanager
-def temp_env(values: dict[str, str]):
-    original = os.environ.copy()
-    os.environ.update(values)
-    try:
-        yield
-    finally:
-        os.environ.clear()
-        os.environ.update(original)
+from tests.outbound_telemetry.conftest import temp_env
 
 
 def test_setup_logging_and_ensure_handler():
